+++ sandbox/cli/admin.js
@@ -38,6 +38,32 @@ function getDailyMsgStat() {
         .finally(() => client.close());
 }
 
+async function _getUsers(client) {
+    await client.connect();
+    const db = client.db(dbName);
+    const c = db.collection('users');
+    const r = await c.find({}).toArray();
+
+    var table = new Table({
+        head: ['Name', 'Email', 'CreatedAt'],
+    });
+
+    r.forEach((item) => {
+        const row = [item.name, item.email, item.createdAt.toISOString()];
+        table.push(row);
+    });
+    console.log(table.toString());
+    return true;
+}
+
+function getUsers() {
+    const client = new MongoClient(url);
+    _getUsers(client)
+        .catch(console.error)
+        .finally(() => client.close());
+}
+
 module.exports = {
     getDailyMsgStat,
+    getUsers,
 }
