+++ sandbox/cli/package.json
@@ -15,6 +15,7 @@
     "commander": "^11.1.0",
     "dotenv": "^16.3.1",
     "eslint": "^9.0.0",
+    "eslint-config-prettier": "^9.1.0",
     "express": "^4.19.2",
     "kill-port": "^2.0.1",
     "mongodb": "^6.3.0",
